azure-functions/archive-old-records/requirements.txt
azure-functions
azure-cosmos
azure-storage-blob
pandas
pyarrow